all forms bugs fixed.

+++ Lab.Application.Contract/Part/CreatePart.cs
@@ -9,6 +9,6 @@ namespace Ex.Application.Contracts.Part
         public Guid PartGroupGuid { get; set; }
         [Required]
         public required string Name { get; set; }
-        public int? StandardWireConsumption { get; set; }
+        public decimal? StandardWireConsumption { get; set; }
     }
 }
